Fixed a bug where the downloads and views are not counted because they relevant code is below the cache line
Created /templates/include.php to be included first in every script. It starts the session so we can track if a person has downloaded/viewed a project already and therefore don't count it.
This can be gamed by closing the browser, but is a quick and easy first fix for the download/view stats